Day6/time_simulator_project/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
openai>=1.3.7
langchain>=0.1.0
langchain-community>=0.0.26
faiss-cpu>=1.7.4
tiktoken>=0.5.1
httpx>=0.25.0
python-dotenv>=1.0.0
streamlit
requests
python-dotenv
langchain_google_genai